backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.4
google-genai==1.7.0
PyJWT==2.8.0
passlib==1.7.4
pytest==8.1.1
Pillow==10.3.0
httpx==0.27.0
python-multipart==0.0.9
